+++ lib/rich_form_field.dart
@@ -9,3 +9,4 @@ part 'src/rich_text_editor_strings.dart';
 part 'src/html_rich_text_form_field.dart';
 part 'src/html_rich_text_controller.dart';
 part 'src/html_rich_text_helpers.dart';
+part 'src/rich_styled_text.dart';